Add choose workingDirectory to preferences
