+++ src/renderer/components/modals/PreferencesModal/ChooseWorkingDirectory/ChooseWorkingDirectory.css
@@ -1,4 +1,4 @@
-.export-directory__name {
+.choose-working-directory__name {
   font-family: monospace;
   word-break: break-word;
 }
