+++ src/renderer/components/modals/PreferencesModal/PreferencesModal.tsx
@@ -18,6 +18,7 @@ import Page from "../../common/Page/Page";
 import PageBody from "../../common/PageBody/PageBody";
 import Toolbar from "../../common/Toolbar/Toolbar";
 import ToolbarItem, { ToolbarItemAlign } from "../../common/ToolbarItem/ToolbarItem";
+import ChooseWorkingDirectory from "./ChooseWorkingDirectory/ChooseWorkingDirectory";
 
 const PreferencesModal = (): JSX.Element => {
   const dispatch = useDispatch();
@@ -41,6 +42,8 @@ const PreferencesModal = (): JSX.Element => {
           <PageBody>
             <Content>
               <ContentBlock title="Preferences">
+                <h3>Interface</h3>
+
                 <InputGroup row>
                   <InputSwitch
                     id="preferencesPlayCaptureSound"
@@ -55,6 +58,25 @@ const PreferencesModal = (): JSX.Element => {
                   />
                   <InputLabel inputId="preferencesPlayCaptureSound">Play capture sound</InputLabel>
                 </InputGroup>
+
+                <InputGroup row>
+                  <InputSwitch
+                    id="preferencesShowTimestampInSeconds"
+                    checked={userPreferences.showTimestampInSeconds}
+                    onChange={() =>
+                      dispatch(
+                        editUserPreferences({
+                          showTimestampInSeconds: !userPreferences.showTimestampInSeconds,
+                        })
+                      )
+                    }
+                  />
+                  <InputLabel inputId="preferencesShowTimestampInSeconds">
+                    Initially show timestamp in seconds rather than frames
+                  </InputLabel>
+                </InputGroup>
+
+                <h3>Playback</h3>
 
                 <InputGroup row>
                   <InputLabel inputId="preferencesShortPlayLength">Short play length</InputLabel>
@@ -74,22 +96,9 @@ const PreferencesModal = (): JSX.Element => {
                   />
                 </InputGroup>
 
-                <InputGroup row>
-                  <InputSwitch
-                    id="preferencesShowTimestampInSeconds"
-                    checked={userPreferences.showTimestampInSeconds}
-                    onChange={() =>
-                      dispatch(
-                        editUserPreferences({
-                          showTimestampInSeconds: !userPreferences.showTimestampInSeconds,
-                        })
-                      )
-                    }
-                  />
-                  <InputLabel inputId="preferencesShowTimestampInSeconds">
-                    Show timestamp in seconds on startup
-                  </InputLabel>
-                </InputGroup>
+                <h3>PROJECTS</h3>
+
+                <ChooseWorkingDirectory />
               </ContentBlock>
             </Content>
           </PageBody>
