+++ src/renderer/redux/thunks.ts
@@ -36,7 +36,7 @@ export const changeWorkingDirectory = (workingDirectory?: string) => {
     return (async () => {
       const newDirectory = await window.preload.ipcToMain.openDirDialog({
         workingDirectory,
-        title: "Select a directory to save captured frames",
+        title: "Select a directory to save projects to",
       });
 
       dispatch(
