+++ src/renderer/styles/theme.css
@@ -51,7 +51,8 @@ h2 {
 }
 
 h3 {
-  font-size: 1.2rem;
+  font-size: 1.1rem;
+  text-transform: uppercase;
 }
 
 input[type="number"],
